+++ test/Arduino.h
@@ -6,9 +6,9 @@
 #include <math.h>
 #include <stdint.h>
 #include <cstring>
+#include "esp32.h"
 
 constexpr auto ARDUINO_PINS = 32;
-constexpr auto LEDC_CHANNELS = 16;
 
 void arduinoMockInit();
 
@@ -24,22 +24,5 @@ void arduinoMockSetMillis(uint32_t value);
 #define PI 3.1415926535897932384626433832795
 #define OUTPUT 0x1
 
-// ESP32 sepcific functions, see
-// packages/framework-arduinoespressif32/cores/esp32/esp32-hal-ledc.h
-double ledcSetup(uint8_t chan, double freq, uint8_t bit_num);
-
-// this struct records calls to ledcSetup
-struct LedcSetupState {
-    double freq;
-    uint8_t bit_num;
-};
-
-struct LedcSetupState arduinoMockGetLedcSetup(uint8_t chan);
-
-void ledcAttachPin(uint8_t pin, uint8_t chan);
-uint8_t arduinoMockGetLedcAttachPin(uint8_t pin);
-
-void ledcWrite(uint8_t chan, uint32_t duty);
-uint32_t arduinoMockGetLedcState(uint8_t chan);
 
 #endif  // TEST_ARDUINO_H_
