+++ test/Makefile
@@ -3,17 +3,23 @@
 CFLAGS=-std=c++14 -c -Wall -Wextra -I. -I../src --coverage -fno-inline \
 	   -fno-inline-small-functions -fno-default-inline -g -fmax-errors=5 \
 	   -fno-omit-frame-pointer -fno-optimize-sibling-calls \
-	   -DHAS_TYPE_TRAITS $(OPT)
+	   $(OPT)
 
 LDFLAGS=-fprofile-arcs -ftest-coverage 
 
-TEST_JLED_SRC=Arduino.cpp test_jled.cpp test_mock.cpp test_main.cpp ../src/jled_base.cpp
+TEST_ARDUINO_MOCK_SRC=Arduino.cpp test_arduino_mock.cpp test_main.cpp 
+TEST_ARDUINO_MOCK_OBJECTS=$(TEST_ARDUINO_MOCK_SRC:.cpp=.o) 
+
+TEST_ESP32_MOCK_SRC=esp32.cpp test_esp32_mock.cpp test_main.cpp 
+TEST_ESP32_MOCK_OBJECTS=$(TEST_ESP32_MOCK_SRC:.cpp=.o) 
+
+TEST_JLED_SRC=Arduino.cpp test_jled.cpp test_main.cpp ../src/jled_base.cpp
 TEST_JLED_OBJECTS=$(TEST_JLED_SRC:.cpp=.o) 
 
-TEST_JLED_SEQUENCE_SRC=Arduino.cpp test_jled_sequence.cpp test_mock.cpp test_main.cpp ../src/jled_base.cpp
+TEST_JLED_SEQUENCE_SRC=Arduino.cpp test_jled_sequence.cpp test_main.cpp ../src/jled_base.cpp
 TEST_JLED_SEQUENCE_OBJECTS=$(TEST_JLED_SEQUENCE_SRC:.cpp=.o)
 
-TEST_ESP32_SRC=Arduino.cpp test_esp32_hal.cpp ../src/esp32_hal.cpp test_main.cpp
+TEST_ESP32_SRC=Arduino.cpp esp32.cpp test_esp32_hal.cpp ../src/esp32_hal.cpp test_main.cpp
 TEST_ESP32_OBJECTS=$(TEST_ESP32_SRC:.cpp=.o)
 
 TEST_ESP8266_SRC=Arduino.cpp test_esp8266_hal.cpp test_main.cpp
@@ -26,7 +32,13 @@ TEST_MORSE_SRC=test_example_morse.cpp test_main.cpp
 TEST_MORSE_OBJECTS=$(TEST_MORSE_SRC:.cpp=.o)
 
 
-all: bin bin/test_jled bin/test_jled_sequence bin/test_esp32_hal bin/test_esp8266_hal bin/test_arduino_hal bin/test_example_morse
+all: bin bin/test_arduino_mock bin/test_esp32_mock bin/test_jled bin/test_jled_sequence bin/test_esp32_hal bin/test_esp8266_hal bin/test_arduino_hal bin/test_example_morse
+
+bin/test_arduino_mock: $(TEST_ARDUINO_MOCK_OBJECTS)
+	$(CXX) -o $@ $(LDFLAGS) $(TEST_ARDUINO_MOCK_OBJECTS) 
+
+bin/test_esp32_mock: $(TEST_ESP32_MOCK_OBJECTS)
+	$(CXX) -o $@ $(LDFLAGS) $(TEST_ESP32_MOCK_OBJECTS) 
 
 bin/test_jled: $(TEST_JLED_OBJECTS)
 	$(CXX) -o $@ $(LDFLAGS) $(TEST_JLED_OBJECTS)
